requirements.txt
farm-haystack
streamlit
streamlit_chat
torch
sentence_transformers
pytube
faiss_cpu
ctransformers
git+https://github.com/openai/whisper.git 
llama-cpp-python